+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@
   - By setting `dynamic_task_generation.enabled: True` in the configuration file, additional tasks (`tasks_per_generation`) can be generated at specified intervals (`interval_seconds`) until the maximum number of generations (`max_generations`) is reached.
   - Various scripts (`agent.py`, `task.py`, `grape.py`) have been updated to be compatible with the integration of these newly-generated tasks during a mission.
 
+- **Rendering Mode**: 
+  - Introduced `agent_work_done` feature: Displays the total distance moved and the amount of work done by each agent.
 
 ## Version 1.2.3 (24.07-19)